requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
beautifulsoup4==4.6.0
bs4==0.0.1
factory-boy==2.8.1
Faker==0.7.12
python-dateutil==2.6.0
six==1.10.0
beautifulsoup4==4.6.0
bs4==0.0.1
factory-boy==2.8.1
Faker==0.7.12
lxml==3.7.3
nose==1.3.7
python-dateutil==2.6.0
six==1.10.0
